requirements-dev.txt
black
flake8
isort
mypy
pre-commit

aws-cdk-lib
aws-cdk-aws-apigatewayv2-alpha==2.61.1a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.61.1a0
aws-solutions-constructs.aws-apigateway-sqs==2.30.0
constructs